rearrang inputs
