+++ src/components/TranslateText.js
@@ -3,6 +3,7 @@ import Languages from './Languages';
 import Input from './Input';
 import Output from './Output';
 import useTranslate from '../hooks/useTranslate';
+import { TranslateButton } from './index';
 
 const TranslateText = () => {
 	const {
@@ -16,28 +17,39 @@ const TranslateText = () => {
 		response,
 		switched,
 		setSwitched,
-    	reset,
+		reset,
 		inputConfig
-		} = useTranslate();
+	} = useTranslate();
+
+	const handleTranslate = () => {
+		translate(input, inputLanguage);
+	  }
 
 	return (
 		<div className="TranslateText">
-			<Languages 
-				setInputLanguage={setInputLanguage} 
-				switched={switched} 
+			<Languages
+				setInputLanguage={setInputLanguage}
+				switched={switched}
 				setSwitched={setSwitched}
-        		reset={reset}
+				reset={reset}
 			/>
-			<Output response={response} loading={loading} inputLanguage={inputLanguage} />
 			<Input
 				input={input}
-				setInput={setInput} 
-				translate={translate} 
-				setLoading={setLoading} 
+				setInput={setInput}
+				translate={translate}
+				setLoading={setLoading}
 				inputLanguage={inputLanguage}
 				inputConfig={inputConfig}
 				switched={switched}
-        		reset={reset}
+				reset={reset}
+			/>
+			<Output response={response} loading={loading} inputLanguage={inputLanguage} />
+			<TranslateButton
+				text={inputConfig.translate}
+				input={input}
+				handleTranslate={handleTranslate}
+				resetText={inputConfig.reset}
+				reset={reset}
 			/>
 		</div>
 	);
