+++ src/styles/Input.css
@@ -59,6 +59,6 @@
     font-size: 16px;
     color: grey;
     position: absolute;
-    margin-top: 12%;
-    right: 48px;
+    margin-top: 38%;
+    right: 15%;
 }
