+++ src/styles/Output.css
@@ -3,7 +3,7 @@
     height: 24vh;
     border-radius: 20px;
     border: none;
-    background: white;
+    background: lightgrey;
     margin-bottom: 32px;
     overflow-y: scroll;
     font-size: 18px;
